Refactored NUnit tests into MS Unit Tests
