+++ .gitignore
@@ -2,6 +2,7 @@
 # This .gitignore file was automatically created by Microsoft(R) Visual Studio.
 ################################################################################
 
+/src/TestResults/*.*
 /src/.vs/*.*
 /src/packages/*.*
 /src/YalvLib.Tests/bin/Debug
@@ -21,3 +22,6 @@
 /src/YALV.Samples/obj/x86/Debug
 /src/YALV.v11.suo
 /src/.vs/YALV
+/src/UnitTests/YalvLib.IntegrationTests/bin/Debug
+/src/UnitTests/YalvLib.IntegrationTests/obj/Debug
+/src/UnitTests/YalvLib.UnitTests/YalvLib.UnitTests/bin/Debug
