--- src/YalvLib.Tests.Integration/Model/sample.xml
@@ -1,13 +1,0 @@
-<log4j:event logger="YALV.Samples.LogService" timestamp="1366474722653" level="ERROR" thread="10">
-  <log4j:message>This is an error message!</log4j:message>
-  <log4j:properties>
-    <log4j:data name="SampleDate" value="20130420" />
-    <log4j:data name="log4japp" value="YALV.Samples.vshost.exe" />
-    <log4j:data name="log4net:UserName" value="tongbong-PC\tongbong" />
-    <log4j:data name="log4jmachinename" value="tongbong-PC" />
-    <log4j:data name="log4net:HostName" value="tongbong-PC" />
-  </log4j:properties>
-  <log4j:throwable>System.Exception: Warning Exception!</log4j:throwable>
-  <log4j:locationInfo class="YALV.Samples.MainWindow" method="method4" file="c:\Workspace\YalvLib\src\YALV.Samples\MainWindow.xaml.cs" line="76" />
-</log4j:event>
-<log4j:event logger="YALV.Samples.LogService" timestamp="1366474722762" level="DEBUG" thread="10"><log4j:message>This is a debug message</log4j:message><log4j:properties><log4j:data name="SampleDate" value="20130420" /><log4j:data name="log4japp" value="YALV.Samples.vshost.exe" /><log4j:data name="log4net:UserName" value="tongbong-PC\tongbong" /><log4j:data name="log4jmachinename" value="tongbong-PC" /><log4j:data name="log4net:HostName" value="tongbong-PC" /></log4j:properties><log4j:locationInfo class="YALV.Samples.MainWindow" method="method1" file="c:\Workspace\YalvLib\src\YALV.Samples\MainWindow.xaml.cs" line="61" /></log4j:event>
