--- src/YalvLib.Tests/Model/TextMarkerTests.cs
@@ -1,70 +1,0 @@
-using System;
-using System.Collections.Generic;
-using NUnit.Framework;
-using YalvLib.Model;
-
-namespace YalvLib.Tests.Model
-{
-    [TestFixture]
-    public class TextMarkerTests
-    {
-        private LogEntry _logEntry;
-        private string _author;
-        private string _message;
-
-        [SetUp]
-        public void CreateEntry()
-        {
-            _logEntry = new LogEntry
-                            {
-                                App = "App",
-                                Class = "Class",
-                                File = "File",
-                                HostName = "Host",
-                                LevelIndex = LevelIndex.ERROR,
-                                Line = 71,
-                                Logger = "Logger",
-                                MachineName = "Machine",
-                                Message = "Message",
-                                Method = "Method",
-                                Thread = "Thread",
-                                Throwable = "Throw",
-                                TimeStamp = DateTime.MaxValue,
-                                UserName = "User"
-                            };
-
-            _author = "Gwen";
-            _message = "Hoo ben gaspard dis pas tout quand meme";
-        }
-
-        [Test]
-        public void CreateTextMarkerTest()
-        {
-            var list = new List<LogEntry> { _logEntry };
-
-            var tm = new TextMarker(list, _author, _message);
-            Assert.AreEqual(tm.Author, _author);
-            Assert.AreEqual(tm.Message, _message);
-            Assert.AreEqual(tm.LogEntries, list);
-            Assert.AreEqual(tm.LogEntryCount(), list.Count);
-            Assert.AreEqual(tm.DateCreation, tm.DateLastModification);
-            Assert.AreEqual(tm.LogEntries, list);
-        }
-
-        [Test]
-        public void EditTextMarkerTest()
-        {
-            var list = new List<LogEntry> { _logEntry };
-
-            var tm = new TextMarker(list, _author, _message)
-                         {Author = "Toto", Message = "On va sur la planete des doudounes quoi!"};
-
-            var dt = new DateTime();
-            tm.DateLastModification = dt;
-
-            Assert.AreEqual(tm.Author, "Toto");
-            Assert.AreEqual(tm.Message, "On va sur la planete des doudounes quoi!");
-            Assert.AreEqual(tm.DateLastModification, dt);
-        }
-    }
-}
